Refactor InspectionSheetForm
